+++ models/index.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 const mongoose = require('mongoose');
-const schema = require('./Schemas/UserSchema');
+const schema = require('./schema/userSchema');
 
 mongoose.connect(process.env.DB_CONNECTION, { useNewUrlParser: true }, err => {
   if (err) return err;
@@ -9,4 +9,4 @@ mongoose.connect(process.env.DB_CONNECTION, { useNewUrlParser: true }, err => {
 
 const User = mongoose.model('Users', schema);
 
-module.exports = { User, mongoose };
+module.exports = { User };
