--- tests/userController.js
@@ -1,40 +1,0 @@
-/* eslint-disable no-undef */
-const chai = require('chai');
-const chaiHttp = require('chai-http');
-const server = require('../index');
-const { User } = require('../models');
-
-chai.should();
-
-chai.use(chaiHttp);
-
-describe('POST /staff/create', () => {
-  const newUser = {
-    username: 'Rexy',
-    password: '1234567'
-  };
-  it('it should add new user account', done => {
-    chai
-      .request(server)
-      .post('/api/v1/staff/create')
-      .send(newUser)
-      .end((err, res) => {
-        res.should.have.status(201);
-        done(err);
-      });
-  });
-  it('it should return 409 ', done => {
-    const newUser = {
-      username: 'Rexy',
-      password: '1234567'
-    };
-    chai
-      .request(server)
-      .post('/api/v1/staff/create')
-      .send(newUser)
-      .end((err, res) => {
-        res.should.have.status(409);
-        done(err);
-      });
-  });
-});
